separating report and graphAPI lambdas
