+++ package.json
@@ -40,6 +40,7 @@
     "@typescript-eslint/parser": "^4.28.0",
     "eslint": "^7.29.0",
     "serverless-add-api-key": "^4.1.1",
+    "serverless-domain-manager": "^6.0.2",
     "serverless-offline": "^7.0.0",
     "serverless-plugin-typescript": "^1.1.9",
     "typescript": "^4.2.4"
